requirements.txt
arxiv
beautifulsoup4
ipykernel
chromadb
langchain
langchain-chroma
ipywidgets
langchain_community
faiss-cpu
langchain-groq
langchain-text-splitters
langchain-openai
python-dotenv
pandas
pypdf
wikipedia
sentence-transformers
langchain_huggingface
